+++ src/request.rs
@@ -1,5 +1,7 @@
 //! Helper functions for request handlers.
 
+use std::io;
+
 use crate::{frame, message};
 use bincode::Options;
 use quinn::SendStream;
@@ -9,12 +11,16 @@ use serde::{Deserialize, Serialize};
 ///
 /// # Errors
 ///
-/// Returns `frame::RecvError::DeserializationFailure`: if the arguments could
-/// not be deserialized.
-pub fn parse_args<'de, T: Deserialize<'de>>(args: &'de [u8]) -> Result<T, frame::RecvError> {
+/// Returns an error if the arguments could not be deserialized.
+pub fn parse_args<'de, T: Deserialize<'de>>(args: &'de [u8]) -> io::Result<T> {
     bincode::DefaultOptions::new()
         .deserialize::<T>(args)
-        .map_err(frame::RecvError::DeserializationFailure)
+        .map_err(|e| {
+            io::Error::new(
+                io::ErrorKind::InvalidData,
+                format!("failed deserializing message: {e}"),
+            )
+        })
 }
 
 /// Sends a response to a request.